Display countrycode and recent geopose on dashboard
