+++ src/core/common.js
@@ -5,9 +5,6 @@
 
 /* Provider for common data types and functions */
 
-import {supportedCountries} from 'ssd-access';
-import * as h3 from "h3-js";
-
 
 /**
  * Type for location info, no orientation
@@ -51,6 +48,13 @@ export const GEOPOSE = {
     quaternion: []
 }
 
+export const LOCALPOSE = {
+    transform: {x: 0, y: 0, z: 0, w: 1},
+    orientation: {x: 0, y: 0, z: 0, w: 0},
+    matrix: {},
+    inverse: {}
+}
+
 /**
  * Empty service value, contained in the services array of an SSR.
  *
@@ -79,54 +83,6 @@ export const ARMODES = {
     marker: 'Marker',
     oscp: 'ARCloud'
 }
-
-
-/**
- *  Promise resolving to the current location (lat, lon) and region code (country currently) of the device.
- *
- * @returns {Promise<LOCATIONINFO>}     Object with lat, lon, regionCode or rejects
- */
-export function getCurrentLocation() {
-    return new Promise((resolve, reject) => {
-        if ('geolocation' in navigator) {
-            navigator.geolocation.getCurrentPosition((position) => {
-                const latAngle = position.coords.latitude;
-                const lonAngle = position.coords.longitude;
-
-                fetch(`https://nominatim.openstreetmap.org/reverse?
-                        lat=${latAngle}&lon=${lonAngle}&format=json&zoom=1&email=info%40michaelvogt.eu`)
-                    .then((response) => {
-                        if (response.ok) {
-                            return response.json();
-                        } else {
-                            reject(response.text());
-                        }
-                    })
-                    .then((data) => {
-                        const countryCode = data.address.country_code;
-                        resolve({
-                            h3Index: h3.geoToH3(latAngle, lonAngle, 8),
-                            lat: latAngle,
-                            lon: lonAngle,
-                            regionCode: supportedCountries.includes(countryCode) ? countryCode : 'us'
-                        })
-                    })
-                    .catch((error) => {
-                        reject(error.statusText());
-                    });
-            }, (error) => {
-                console.log(`Location request failed: ${error}`)
-                reject(error);
-            }, {
-                enableHighAccuracy: false,
-                maximumAge: 0
-            });
-        } else {
-            reject('Location is not available');
-        }
-    });
-}
-
 
 /**
  * Converting a WebGLTexture to base64 encoded image.
