+++ src/core/locationTools.js
@@ -9,7 +9,8 @@
 
 import LatLon from 'geodesy/latlon-ellipsoidal-vincenty.js';
 import { quat, vec3 } from 'gl-matrix';
-
+import * as h3 from "h3-js";
+import { supportedCountries } from 'ssd-access';
 
 export const toRadians = (degrees) => degrees * Math.PI / 180;
 export const toDegrees = (radians) => radians / Math.PI / 180;
@@ -149,6 +150,53 @@ export const fakeLocationResult = {
     ]
 }
 
+
+/**
+ *  Promise resolving to the current location (lat, lon) and region code (country currently) of the device.
+ *
+ * @returns {Promise<LOCATIONINFO>}     Object with lat, lon, regionCode or rejects
+ */
+export function getCurrentLocation() {
+    return new Promise((resolve, reject) => {
+        if ('geolocation' in navigator) {
+            navigator.geolocation.getCurrentPosition((position) => {
+                const latAngle = position.coords.latitude;
+                const lonAngle = position.coords.longitude;
+
+                fetch(`https://nominatim.openstreetmap.org/reverse?
+                        lat=${latAngle}&lon=${lonAngle}&format=json&zoom=1&email=info%40michaelvogt.eu`)
+                    .then((response) => {
+                        if (response.ok) {
+                            return response.json();
+                        } else {
+                            reject(response.text());
+                        }
+                    })
+                    .then((data) => {
+                        const countryCode = data.address.country_code;
+                        resolve({
+                            h3Index: h3.geoToH3(latAngle, lonAngle, 8),
+                            lat: latAngle,
+                            lon: lonAngle,
+                            countryCode: countryCode,
+                            regionCode: supportedCountries.includes(countryCode) ? countryCode : 'us'
+                        })
+                    })
+                    .catch((error) => {
+                        reject(error.statusText());
+                    });
+            }, (error) => {
+                console.log(`Location request failed: ${error}`)
+                reject(error);
+            }, {
+                enableHighAccuracy: false,
+                maximumAge: 0
+            });
+        } else {
+            reject('Location is not available');
+        }
+    });
+}
 
 /**
  *  Calculates the relative distance of two geodesic locations.
